requirements.txt
requests
gemfileparser2
pyyaml
requirements-parser